# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Clinic Management System API",
    "description": "API documentation for the Patient Mobile Application",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token in format: Bearer <token>"
    }
  },
  "tags": [
    {
      "name": "Authentication",
      "description": "Authentication endpoints"
    },
    {
      "name": "Profile",
      "description": "Patient profile management"
    },
    {
      "name": "Appointments",
      "description": "Appointment management"
    },
    {
      "name": "Cases",
      "description": "Medical cases management"
    },
    {
      "name": "Questions",
      "description": "Doctor-patient communication"
    },
    {
      "name": "Feedback",
      "description": "Doctor feedback management"
    }
  ],
  "paths": {
    "/patient/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Patient login",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "password"],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "patient@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "patient_number": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/patient/profile": {
      "get": {
        "tags": ["Profile"],
        "summary": "Get patient profile",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Patient profile data",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "patient_number": {
                  "type": "string"
                },
                "mobile_number": {
                  "type": "string"
                },
                "gender": {
                  "type": "string"
                },
                "date_of_birth": {
                  "type": "string",
                  "format": "date"
                },
                "address": {
                  "type": "string"
                },
                "city": {
                  "type": "string"
                },
                "state": {
                  "type": "string"
                },
                "pin_code": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/patient/appointments": {
      "get": {
        "tags": ["Appointments"],
        "summary": "Get patient appointments",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "type": "string",
            "enum": ["scheduled", "completed", "cancelled"],
            "description": "Filter appointments by status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of appointments",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "datetime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "status": {
                    "type": "string"
                  },
                  "doctor": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patient/cases": {
      "get": {
        "tags": ["Cases"],
        "summary": "Get patient cases",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of cases",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "created_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "doctor": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  },
                  "diagnosis": {
                    "type": "string"
                  },
                  "treatment": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patient/questions": {
      "get": {
        "tags": ["Questions"],
        "summary": "Get patient questions",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of questions",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "question": {
                    "type": "string"
                  },
                  "answer": {
                    "type": "string"
                  },
                  "created_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "answered_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "doctor": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Questions"],
        "summary": "Create a new question",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["doctor_id", "question"],
              "properties": {
                "doctor_id": {
                  "type": "integer"
                },
                "question": {
                  "type": "string"
                },
                "is_private": {
                  "type": "boolean",
                  "default": false
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Question created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "id": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/patient/feedback": {
      "post": {
        "tags": ["Feedback"],
        "summary": "Submit doctor feedback",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["doctor_id", "rating"],
              "properties": {
                "doctor_id": {
                  "type": "integer"
                },
                "rating": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 5
                },
                "comment": {
                  "type": "string"
                },
                "is_anonymous": {
                  "type": "boolean",
                  "default": false
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Feedback submitted successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    }
  }
}